Completed Hardcode and working Controller
A raw working system is now made. This system uses a hardcoded
questionnaire. The newer OO one will be implemented later on. This is
basically what rick was asking for and all we are now missing is the
pickle, COMSC admin and the matplotlib. With this code we technically
are close to completion. Each of the modules have test buttons which is
to show how the modules are independent. Don’t change anything on these
modules as they are completed. If you run the controller, you can use
the questionnaire.